+++ src/components/stable/gux-action-toast/tests/gux-action-toast.spec.ts
@@ -11,10 +11,10 @@ describe('gux-action-toast', () => {
         description: 'should render action toast',
         html: `
           <gux-action-toast accent="neutral">
-            <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+            <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
             <div slot="title">Title</div>
             <div slot="message">This is the message</div>
-            <gux-button slot="negitive-button">Reject</gux-button>
+            <gux-button slot="negative-button">Reject</gux-button>
             <gux-button slot="positive-button" accent="primary">Accept</gux-button>
           </gux-action-toast>
         `