+++ src/components/stable/gux-advanced-dropdown/dropdown-option/gux-dropdown-option.tsx
@@ -16,6 +16,7 @@ function escapeRegex(input) {
 }
 
 @Component({
+  styleUrl: 'gux-dropdown-option.less',
   tag: 'gux-dropdown-option'
 })
 export class GuxDropdownOption {
@@ -43,13 +44,13 @@ export class GuxDropdownOption {
    *
    * Should only be used by internal users.
    */
-  @Prop({ mutable: true, reflectToAttr: true })
+  @Prop({ mutable: true, reflect: true })
   filtered: boolean;
 
   /**
    * If present, this Boolean attribute indicates that the option is currently selected.
    */
-  @Prop({ mutable: true, reflectToAttr: true })
+  @Prop({ mutable: true, reflect: true })
   selected: boolean;
 
   @Prop()