+++ src/components/stable/gux-datepicker/gux-datepicker.tsx
@@ -616,7 +616,7 @@ export class GuxDatepicker {
                 }}
                 tabindex="-1"
               >
-                <gux-icon decorative iconName="ic-calendar-generic"></gux-icon>
+                <gux-icon decorative icon-name="calendar"></gux-icon>
               </button>
               <gux-calendar
                 ref={(el: HTMLGuxCalendarElement) =>
@@ -657,10 +657,7 @@ export class GuxDatepicker {
                   }}
                   tabindex="-1"
                 >
-                  <gux-icon
-                    decorative
-                    iconName="ic-calendar-generic"
-                  ></gux-icon>
+                  <gux-icon decorative icon-name="calendar"></gux-icon>
                 </button>
               </gux-text-field-legacy>
             </gux-text-label>