+++ src/components/stable/gux-form-field/components/gux-input-text-like/tests/gux-input-text-like.spec.ts
@@ -19,7 +19,7 @@ describe('gux-input-text-like', () => {
           <input type="text" slot="input"/>
         </gux-input-text-like>
       `,
-      language: 'en'
+      language
     });
   });