+++ src/components/stable/gux-notification-toast/tests/gux-notification-toast.e2e.ts
@@ -7,7 +7,7 @@ describe('gux-notification-toast', () => {
         description: 'should render neutral notification toast',
         html: `
           <gux-notification-toast lang="en" accent="neutral">
-            <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+            <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
             <div slot="title">Title</div>
             <div slot="message">This is the message</div>
           </gux-notification-toast>
@@ -17,7 +17,7 @@ describe('gux-notification-toast', () => {
         description: 'should render positive notification toast',
         html: `
           <gux-notification-toast lang="en" accent="positive">
-            <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+            <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
             <div slot="title">Title</div>
             <div slot="message">This is the message</div>
           </gux-notification-toast>
@@ -27,7 +27,7 @@ describe('gux-notification-toast', () => {
         description: 'should render alert notification toast',
         html: `
           <gux-notification-toast lang="en" accent="alert">
-            <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+            <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
             <div slot="title">Title</div>
             <div slot="message">This is the message</div>
           </gux-notification-toast>
@@ -37,7 +37,7 @@ describe('gux-notification-toast', () => {
         description: 'should render warning notification toast',
         html: `
           <gux-notification-toast lang="en" accent="warning">
-            <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+            <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
             <div slot="title">Title</div>
             <div slot="message">This is the message</div>
           </gux-notification-toast>
@@ -57,7 +57,7 @@ describe('gux-notification-toast', () => {
     it('click dismiss button', async () => {
       const html = `
         <gux-notification-toast lang="en" accent="neutral">
-          <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+          <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
           <div slot="title">Title</div>
           <div slot="message">This is the message</div>
         </gux-notification-toast>
@@ -82,7 +82,7 @@ describe('gux-notification-toast', () => {
     it('click dismiss button and prevent default', async () => {
       const html = `
         <gux-notification-toast lang="en" accent="neutral">
-          <gux-icon slot="icon" icon-name="add-user" decorative></gux-icon>
+          <gux-icon slot="icon" icon-name="user-add" decorative></gux-icon>
           <div slot="title">Title</div>
           <div slot="message">This is the message</div>
         </gux-notification-toast>